# Improve performance of `hashObject`