--- a/lib/counter.js
+++ b/lib/counter.js
@@ -40,7 +40,7 @@ class Counter extends Metric {
 	incWithoutExemplar(labels, value) {
 		let hash = '';
 		if (isObject(labels)) {
-			hash = hashObject(labels);
+			hash = hashObject(labels, this.sortedLabelNames);
 			validateLabel(this.labelNames, labels);
 		} else {
 			value = labels;
@@ -130,7 +130,7 @@ class Counter extends Metric {
 	remove(...args) {
 		const labels = getLabels(this.labelNames, args) || {};
 		validateLabel(this.labelNames, labels);
-		return removeLabels.call(this, this.hashMap, labels);
+		return removeLabels.call(this, this.hashMap, labels, this.sortedLabelNames);
 	}
 }