--- a/lib/gauge.js
+++ b/lib/gauge.js
@@ -120,7 +120,7 @@ class Gauge extends Metric {
 	}
 
 	_getValue(labels) {
-		const hash = hashObject(labels || {});
+		const hash = hashObject(labels || {}, this.sortedLabelNames);
 		return this.hashMap[hash] ? this.hashMap[hash].value : 0;
 	}
 
@@ -139,7 +139,7 @@ class Gauge extends Metric {
 	remove(...args) {
 		const labels = getLabels(this.labelNames, args);
 		validateLabel(this.labelNames, labels);
-		removeLabels.call(this, this.hashMap, labels);
+		removeLabels.call(this, this.hashMap, labels, this.sortedLabelNames);
 	}
 }
 
@@ -158,7 +158,7 @@ function setDelta(gauge, labels, delta) {
 	}
 
 	validateLabel(gauge.labelNames, labels);
-	const hash = hashObject(labels);
+	const hash = hashObject(labels, gauge.sortedLabelNames);
 	setValueDelta(gauge.hashMap, delta, labels, hash);
 }