--- a/lib/histogram.js
+++ b/lib/histogram.js
@@ -83,7 +83,7 @@ class Histogram extends Metric {
 	}
 
 	updateExemplar(labels, value, exemplarLabels) {
-		const hash = hashObject(labels);
+		const hash = hashObject(labels, this.sortedLabelNames);
 		const bound = findBound(this.upperBounds, value);
 		const { bucketExemplars } = this.hashMap[hash];
 		let exemplar = bucketExemplars[bound];
@@ -132,7 +132,7 @@ class Histogram extends Metric {
 	 * @returns {void}
 	 */
 	zero(labels) {
-		const hash = hashObject(labels);
+		const hash = hashObject(labels, this.sortedLabelNames);
 		this.hashMap[hash] = createBaseValues(
 			labels,
 			this.bucketValues,
@@ -167,7 +167,7 @@ class Histogram extends Metric {
 	remove(...args) {
 		const labels = getLabels(this.labelNames, args);
 		validateLabel(this.labelNames, labels);
-		removeLabels.call(this, this.hashMap, labels);
+		removeLabels.call(this, this.hashMap, labels, this.sortedLabelNames);
 	}
 }
 
@@ -214,7 +214,7 @@ function observe(labels) {
 			);
 		}
 
-		const hash = hashObject(labelValuePair.labels);
+		const hash = hashObject(labelValuePair.labels, this.sortedLabelNames);
 		let valueFromMap = this.hashMap[hash];
 		if (!valueFromMap) {
 			valueFromMap = createBaseValues(