--- a/lib/metric.js
+++ b/lib/metric.js
@@ -39,8 +39,15 @@ class Metric {
 		if (!validateLabelName(this.labelNames)) {
 			throw new Error('Invalid label name');
 		}
+
 		if (this.collect && typeof this.collect !== 'function') {
 			throw new Error('Optional "collect" parameter must be a function');
+		}
+
+		if (this.labelNames) {
+			this.sortedLabelNames = [...this.labelNames].sort();
+		} else {
+			this.sortedLabelNames = [];
 		}
 
 		this.reset();