--- a/lib/summary.js
+++ b/lib/summary.js
@@ -115,7 +115,7 @@ class Summary extends Metric {
 	remove(...args) {
 		const labels = getLabels(this.labelNames, args);
 		validateLabel(this.labelNames, labels);
-		removeLabels.call(this, this.hashMap, labels);
+		removeLabels.call(this, this.hashMap, labels, this.sortedLabelNames);
 	}
 }
 
@@ -170,7 +170,7 @@ function observe(labels) {
 			);
 		}
 
-		const hash = hashObject(labelValuePair.labels);
+		const hash = hashObject(labelValuePair.labels, this.sortedLabelNames);
 		let summaryOfLabel = this.hashMap[hash];
 		if (!summaryOfLabel) {
 			summaryOfLabel = {