--- a/lib/util.js
+++ b/lib/util.js
@@ -14,8 +14,12 @@ exports.getValueAsString = function getValueString(value) {
 	}
 };
 
-exports.removeLabels = function removeLabels(hashMap, labels) {
-	const hash = hashObject(labels);
+exports.removeLabels = function removeLabels(
+	hashMap,
+	labels,
+	sortedLabelNames,
+) {
+	const hash = hashObject(labels, sortedLabelNames);
 	delete hashMap[hash];
 };
 
@@ -59,32 +63,44 @@ exports.getLabels = function (labelNames, args) {
 	return acc;
 };
 
-function hashObject(labels) {
-	// We don't actually need a hash here. We just need a string that
-	// is unique for each possible labels object and consistent across
-	// calls with equivalent labels objects.
-	let keys = Object.keys(labels);
+function fastHashObject(keys, labels) {
 	if (keys.length === 0) {
 		return '';
 	}
-	// else
+
+	let hash = '';
+
+	for (let i = 0; i < keys.length; i++) {
+		const key = keys[i];
+		const value = labels[key];
+		if (typeof value === 'undefined') continue;
+
+		hash += `${key}:${value},`;
+	}
+
+	return hash;
+}
+
+function hashObject(labels, labelNames) {
+	// We don't actually need a hash here. We just need a string that
+	// is unique for each possible labels object and consistent across
+	// calls with equivalent labels objects.
+
+	if (labelNames) {
+		return fastHashObject(labelNames, labels);
+	}
+
+	let keys = Object.keys(labels);
 	if (keys.length > 1) {
 		keys = keys.sort(); // need consistency across calls
 	}
 
-	let hash = '';
-	let i = 0;
-	const size = keys.length;
-	for (; i < size - 1; i++) {
-		hash += `${keys[i]}:${labels[keys[i]]},`;
-	}
-	hash += `${keys[i]}:${labels[keys[i]]}`;
-	return hash;
+	return fastHashObject(keys, labels);
 }
 exports.hashObject = hashObject;
 
 exports.isObject = function isObject(obj) {
-	return obj === Object(obj);
+	return obj !== null && typeof obj === 'object';
 };
 
 exports.nowTimestamp = function nowTimestamp() {