--- a/lib/util.js
+++ b/lib/util.js
@@ -90,9 +90,9 @@ function hashObject(labels, labelNames) {
 		return fastHashObject(labelNames, labels);
 	}
 
-	let keys = Object.keys(labels);
+	const keys = Object.keys(labels);
 	if (keys.length > 1) {
-		keys = keys.sort(); // need consistency across calls
+		keys.sort(); // need consistency across calls
 	}
 
 	return fastHashObject(keys, labels);