--- a/lib/util.js
+++ b/lib/util.js
@@ -73,7 +73,7 @@ function fastHashObject(keys, labels) {
 	for (let i = 0; i < keys.length; i++) {
 		const key = keys[i];
 		const value = labels[key];
-		if (typeof value === 'undefined') continue;
+		if (value === undefined) continue;
 
 		hash += `${key}:${value},`;
 	}