@zbjornson 2023-10-25
Thanks for the PR. This has been proposed before in #373 and would make it impossible to implement #298/#368. See also this comment about validly accepting subsets of labels: https://github.com/siimon/prom-client/pull/373#discussion_r426204193.