@zbjornson 2023-10-26
> if I don't get `labelNames` I fallback to extracting them like the old code.

My bad, I missed the fallback. That's a great solution then.

Can you add a changelog entry please?