@yosiat 2023-10-26
@zbjornson added changelog entry, let me know if it's ok.