@SimenB 2023-12-15
https://github.com/siimon/prom-client/releases/tag/v15.1.0